requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
spacy
nltk
pyspellchecker
contractions
